Move App from index to seperate file
